Correção de relacionamentos

+++ Application/Commands/DeleteOrder/DeleteOrderCommandHandler.cs
@@ -1,24 +1,33 @@
 ﻿using MediatR;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Core.Repositories;
 
 namespace Application.Commands.DeleteOrder
 {
-    public class DeleteOrdersCommandHandler : IRequestHandler<DeleteOrderCommand, Unit>
+    public class DeleteOrderCommandHandler : IRequestHandler<DeleteOrderCommand, Unit>
     {
         private readonly IOrderRepository _orderRepository;
-        public DeleteOrdersCommandHandler(IOrderRepository orderRepository)
+        private readonly IOrderProductRepository _orderProductRepository;
+        public DeleteOrderCommandHandler(IOrderRepository orderRepository, IOrderProductRepository orderProductRepository)
         {
             _orderRepository = orderRepository;
+            _orderProductRepository = orderProductRepository;
         }
 
         public async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
         {
             var order = await _orderRepository.GetByIdAsync(request.Id);
+
+            if (order == null)
+            {
+                throw new InvalidOperationException("Order not found.");
+            }
+
+            foreach (var orderProduct in order.OrderProducts)
+            {
+                await _orderProductRepository.DeleteOrderProductAsync(order.Id, orderProduct.ProductId);
+            }
 
             await _orderRepository.DeleteAsync(request.Id);
