+++ Application/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,8 +1,5 @@
 ﻿using MediatR;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using Core.Repositories;
 
@@ -11,14 +8,26 @@ namespace Application.Commands.DeleteProduct
     public class DeleteProductCommandHandler : IRequestHandler<DeleteProductCommand, Unit>
     {
         private readonly IProductRepository _productRepository;
-        public DeleteProductCommandHandler(IProductRepository productRepository)
+        private readonly IProductRawMaterialRepository _productRawMaterialRepository;
+        public DeleteProductCommandHandler(IProductRepository productRepository, IProductRawMaterialRepository productRawMaterialRepository)
         {
             _productRepository = productRepository;
+            _productRawMaterialRepository = productRawMaterialRepository;
         }
 
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
             var product = await _productRepository.GetByIdAsync(request.Id);
+
+            if (product == null)
+            {
+                throw new InvalidOperationException("Product not found.");
+            }
+
+            foreach (var productRawMaterial in product.ProductRawMaterials)
+            {
+                await _productRawMaterialRepository.DeleteProductRawMaterialAsync(product.Id, productRawMaterial.RawMaterialId);
+            }
 
             await _productRepository.DeleteAsync(request.Id);
